Prevent infinite loop in datachanel open re-transmission
In some cases it is possible for re-transmission code to loop infinitely
when a data channel is connected. This is caused by the fact that onopen
can be called while the datachanel is still in "connecting" readyState.
At this moment the sendDirectly function will queue the message again
and the loop will try to re-transmit the message again on the next
iteration when it will be queue again.

To prevent this infinite loop add a check that the readyState is "open"
so the code don't loop for ever.

Signed-off-by: Vladislav Hristov <vvtthristov@gmail.com>